final fix before send

+++ api/Database/Controllers/AdminAccountController.cs
@@ -14,15 +14,21 @@ namespace api.Database.Controllers
     public class AdminAccountController : ControllerBase
     {
         SimbirGoContext db;
+
         public AdminAccountController(SimbirGoContext db)
         {
             this.db = db;
+                    
         }
         [HttpGet("/Admin/Account")]
         public IActionResult GetAllUsers(int start, int count)
         {
             try
-            {               
+            {
+                var jwt = Request.Headers["Authorization"].ToString().Replace("Bearer ", "");
+                User? user = db.Users.FirstOrDefault(x => x.Username == JwtActions.ReturnUsername(jwt));
+                if (user == null)
+                    return NotFound("Uncorrect token");
                 var users = db.Users.Skip(start).Take(count).ToList();
                 return Ok(users);
             }
@@ -39,9 +45,13 @@ namespace api.Database.Controllers
         {
             try
             {
+                var jwt = Request.Headers["Authorization"].ToString().Replace("Bearer ", "");
+                User? usr = db.Users.FirstOrDefault(x => x.Username == JwtActions.ReturnUsername(jwt));
+                if (usr == null)
+                    return NotFound("Uncorrect token");
                 User? user = db.Users.FirstOrDefault(x => x.Id == id);
                 if (user == null)
-                    return BadRequest("Id not exist");
+                    return NotFound("Id not exist");
                 return Ok(user);
             }
             catch (Exception ex)
@@ -55,7 +65,11 @@ namespace api.Database.Controllers
         {
             try
             {
-                if(db.Users.FirstOrDefault(x=> x.Username == user.Username) == null )
+                var jwt = Request.Headers["Authorization"].ToString().Replace("Bearer ", "");
+                User? usr = db.Users.FirstOrDefault(x => x.Username == JwtActions.ReturnUsername(jwt));
+                if (usr == null)
+                    return NotFound("Uncorrect token");
+                if (db.Users.FirstOrDefault(x=> x.Username == user.Username) == null )
                 {
                     db.Add(user);
                     db.SaveChanges();
@@ -75,9 +89,13 @@ namespace api.Database.Controllers
         {
             try
             {
+                var jwt = Request.Headers["Authorization"].ToString().Replace("Bearer ", "");
+                User? usr = db.Users.FirstOrDefault(x => x.Username == JwtActions.ReturnUsername(jwt));
+                if (usr == null)
+                    return NotFound("Uncorrect token");
                 User? existUser = db.Users.FirstOrDefault(x => x.Id == id);
                 if (existUser == null)
-                    return BadRequest("User with this id not exist");                
+                    return NotFound("User with this id not exist");                
                 if (db.Users.First(x => x.Username == user.Username) == null || existUser.Username == user.Username)
                 {
                     existUser.Username = user.Username;
@@ -103,9 +121,13 @@ namespace api.Database.Controllers
         {
             try
             {
+                var jwt = Request.Headers["Authorization"].ToString().Replace("Bearer ", "");
+                User? usr = db.Users.FirstOrDefault(x => x.Username == JwtActions.ReturnUsername(jwt));
+                if (usr == null)
+                    return NotFound("Uncorrect token");
                 User? user = db.Users.FirstOrDefault(x => x.Id == id);
                 if (user == null)
-                    return BadRequest("Id not exist");
+                    return NotFound("Id not exist");
                 db.Remove(user);
                 db.SaveChanges();
                 return Ok();
