+++ api/Database/Controllers/AdminRentController.cs
@@ -23,9 +23,13 @@ namespace api.Database.Controllers
         {
             try
             {
+                var jwt = Request.Headers["Authorization"].ToString().Replace("Bearer ", "");
+                User? usr = db.Users.FirstOrDefault(x => x.Username == JwtActions.ReturnUsername(jwt));
+                if (usr == null)
+                    return NotFound("Uncorrect token");
                 RentInfo? rent = db.RentInfos.FirstOrDefault(x => x.Id == rentId);
                 if (rent == null)
-                    return BadRequest("Id not exist");
+                    return NotFound("Id not exist");
                 return Ok(rent);
             }
             catch (Exception ex)
@@ -39,9 +43,13 @@ namespace api.Database.Controllers
         {
             try
             {
+                var jwt = Request.Headers["Authorization"].ToString().Replace("Bearer ", "");
+                User? usr = db.Users.FirstOrDefault(x => x.Username == JwtActions.ReturnUsername(jwt));
+                if (usr == null)
+                    return NotFound("Uncorrect token");
                 User? user = db.Users.FirstOrDefault(x => x.Id == userId);
                 if(user == null)
-                    return BadRequest("User with this Id not exist");
+                    return NotFound("User with this Id not exist");
                 var rent = db.RentInfos.Where(x => x.UserId == user.Id).ToList();
                 return Ok(rent);
             }
@@ -56,9 +64,13 @@ namespace api.Database.Controllers
         {
             try
             {
+                var jwt = Request.Headers["Authorization"].ToString().Replace("Bearer ", "");
+                User? usr = db.Users.FirstOrDefault(x => x.Username == JwtActions.ReturnUsername(jwt));
+                if (usr == null)
+                    return NotFound("Uncorrect token");
                 Transport? transport = db.Transports.FirstOrDefault(x => x.Id == transportId);
                 if (transport == null)
-                    return BadRequest("Transport with this Id not exist");
+                    return NotFound("Transport with this Id not exist");
                 var rent = db.RentInfos.Where(x => x.TransportId == transportId).ToList();
                 return Ok(rent);
             }
@@ -72,15 +84,19 @@ namespace api.Database.Controllers
         {
             try
             {
+                var jwt = Request.Headers["Authorization"].ToString().Replace("Bearer ", "");
+                User? usr = db.Users.FirstOrDefault(x => x.Username == JwtActions.ReturnUsername(jwt));
+                if (usr == null)
+                    return NotFound("Uncorrect token");
                 User? user = db.Users.FirstOrDefault(x=> x.Id == rent.UserId);
                 Transport? transport = db.Transports.FirstOrDefault(x => x.Id == rent.TransportId);                
                 if (user == null)
-                    return BadRequest("User with this ID does not exist");
+                    return NotFound("User with this ID does not exist");
                 if(transport == null || transport.CanBeRented == false)
-                    return BadRequest("Transport with this ID does not exist");
+                    return NotFound("Transport with this ID does not exist");
                 RentType? priceType = db.RentTypes.FirstOrDefault(x => x.RentType1 == rent.PriceType);
                 if (priceType == null)
-                    return BadRequest("This type does not exist");
+                    return NotFound("This type does not exist");
                 db.Add(new Rent
                 {                    
                     IdTransport = Convert.ToInt64(transport.Id),
@@ -105,18 +121,22 @@ namespace api.Database.Controllers
         {
             try
             {
+                var jwt = Request.Headers["Authorization"].ToString().Replace("Bearer ", "");
+                User? usr = db.Users.FirstOrDefault(x => x.Username == JwtActions.ReturnUsername(jwt));
+                if (usr == null)
+                    return NotFound("Uncorrect token");
                 Rent? rent = db.Rents.FirstOrDefault(x => x.Id == id);
                 User? user = db.Users.FirstOrDefault(x => x.Id == rentInfo.UserId);
                 Transport? transport = db.Transports.FirstOrDefault(x => x.Id == rentInfo.TransportId);
                 if (rent == null)
-                    return BadRequest("Rent with this ID does not exist");
+                    return NotFound("Rent with this ID does not exist");
                 if (user == null)
-                    return BadRequest("User with this ID does not exist");
+                    return NotFound("User with this ID does not exist");
                 if (transport == null)
-                    return BadRequest("Transport with this ID does not exist");
+                    return NotFound("Transport with this ID does not exist");
                 RentType? priceType = db.RentTypes.FirstOrDefault(x => x.RentType1 == rentInfo.PriceType);
                 if (priceType == null)
-                    return BadRequest("this PriceType with this ID does not exist");
+                    return NotFound("this PriceType with this ID does not exist");
                 rent.IdTransport = Convert.ToInt64(transport.Id);
                 rent.IdUser = Convert.ToInt64(user.Id);
                 rent.TimeStart = Convert.ToDateTime(rentInfo.TimeStart);
@@ -139,9 +159,13 @@ namespace api.Database.Controllers
         {
             try
             {
+                var jwt = Request.Headers["Authorization"].ToString().Replace("Bearer ", "");
+                User? usr = db.Users.FirstOrDefault(x => x.Username == JwtActions.ReturnUsername(jwt));
+                if (usr == null)
+                    return NotFound("Uncorrect token");
                 Rent? rent = db.Rents.FirstOrDefault(x => x.Id == rentId);
                 if (rent == null)
-                    return BadRequest("Id not exist");
+                    return NotFound("Id not exist");
                 db.Remove(rent);
                 db.SaveChanges();
                 return Ok();
@@ -157,11 +181,14 @@ namespace api.Database.Controllers
         {
             try
             {
+                var jwt = Request.Headers["Authorization"].ToString().Replace("Bearer ", "");
+                User? usr = db.Users.FirstOrDefault(x => x.Username == JwtActions.ReturnUsername(jwt));
+                if (usr == null)
+                    return NotFound("Uncorrect token");
                 Rent? rent = db.Rents.FirstOrDefault(x => x.Id == rentId);
                 if (rent != null)
                 {
-                    Transport transport = db.Transports.First(x => x.Id == rent.IdTransport);
-                    
+                    Transport transport = db.Transports.First(x => x.Id == rent.IdTransport);                    
                     rent.TimeEnd = DateTime.UtcNow;
                     TimeSpan difference = DateTime.UtcNow - rent.TimeStart;
                     if (rent.PriceType == 2)
@@ -179,7 +206,7 @@ namespace api.Database.Controllers
                     db.SaveChanges();
                     return Ok();
                 }
-                return BadRequest("Rent with this Id does not exist");
+                return NotFound("Rent with this Id does not exist");
             }
             catch (Exception ex)
             {
